Change photo to component
This will provide photo as a component.
